requirements.txt
Django
djangorestframework
psycopg2
django-rest-swagger
xmljson
lxml